refactor: refactors the animal entity package
