+++ src/main/kotlin/tech/cordona/zooonline/domain/animal/entity/Animal.kt
@@ -24,9 +24,3 @@ data class Animal(
 	override var modifiedDate: LocalDateTime = LocalDateTime.now()
 
 ) : BaseEntity
-
-enum class Gender(var asString: String) {
-	MALE("Male"),
-	FEMALE("Female")
-}
-
