+++ src/main/kotlin/tech/cordona/zooonline/domain/animal/entity/enums/Gender.kt
@@ -1,0 +1,6 @@
+package tech.cordona.zooonline.domain.animal.entity.enums
+
+enum class Gender(var asString: String) {
+	MALE("Male"),
+	FEMALE("Female")
+}
