+++ src/main/kotlin/tech/cordona/zooonline/domain/animal/entity/enums/HealthStatus.kt
@@ -1,0 +1,6 @@
+package tech.cordona.zooonline.domain.animal.entity.enums
+
+enum class HealthStatus(val asString: String) {
+	HEALTHY("Healthy"),
+	SICK("Sick")
+}
