+++ src/main/kotlin/tech/cordona/zooonline/domain/animal/entity/enums/TrainingStatus.kt
@@ -1,0 +1,6 @@
+package tech.cordona.zooonline.domain.animal.entity.enums
+
+enum class TrainingStatus(val asString: String) {
+	TRAINED("Trained"),
+	UNTRAINED("Untrained")
+}
