+++ src/main/kotlin/tech/cordona/zooonline/domain/animal/entity/structs/HealthStatistics.kt
@@ -1,21 +1,11 @@
 package tech.cordona.zooonline.domain.animal.entity.structs
 
 data class HealthStatistics(
-	var healthPoints: Int,
-	var healthStatus: String,
-	var trainingPoints: Int,
-	var trainingStatus: String
-) {
-	enum class TrainingStatus(val asString: String) {
-		TRAINED("Trained"),
-		UNTRAINED("Untrained")
-	}
-
-	enum class HealthStatus(val asString: String) {
-		HEALTHY("Healthy"),
-		SICK("Sick")
-	}
-}
+	val healthPoints: Int,
+	val healthStatus: String,
+	val trainingPoints: Int,
+	val trainingStatus: String
+)
